+++ utils/add_china_db_platform.py
@@ -1,0 +1,8 @@
+#!/usr/bin/python
+#
+databases = [
+    {
+        'name': 'TiDB',
+        'type': 'mysql',
+    }
+]
